requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# ── Streamlit apps ─────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0
gtts>=2.5.0

# ── AI / ML (optional — apps degrade gracefully if missing) ────────────────────
openai>=1.30.0
sentence-transformers>=2.7.0

# ── Optional integrations ──────────────────────────────────────────────────────
mindee>=4.0.0          # Resume parsing (Mindee OCR)
firecrawl-py>=0.0.16   # Candidate link scraping
pdfplumber>=0.11.0     # PDF hyperlink extraction
pymupdf>=1.23.0        # Required for pdf_service via fitz